airbyte-integrations/connectors/source-surveycto/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
version = "0.1.0"
name = "source-surveycto"
description = "Source implementation for surveycto."
authors = [ "Airbyte <contact@airbyte.io>",]
license = "MIT"
readme = "README.md"
documentation = "https://docs.airbyte.com/integrations/sources/surveycto"
homepage = "https://airbyte.com"
repository = "https://github.com/airbytehq/airbyte"
packages = [ { include = "source_surveycto" }, {include = "main.py" } ]

[tool.poetry.dependencies]
python = "^3.10,<3.12"
airbyte-cdk = "^6.13.0"

[tool.poetry.scripts]
source-surveycto = "source_surveycto.run:run"

[tool.poetry.group.dev.dependencies]
requests-mock = "*"
pytest-mock = "*"
pytest = "*"
